Move type/platform-based repo fetching to dedicated modules
